# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "swagger 接口文档",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/ChangSZ/mall-go/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/login": {
            "post": {
                "description": "登录以后返回token",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UmsAdminController"
                ],
                "summary": "登录以后返回token",
                "parameters": [
                    {
                        "description": "请求信息",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_admin.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ums_admin.loginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
        "/admin/register": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UmsAdminController"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "请求信息",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_admin.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ums_admin.registerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
        "/admin/{id}": {
            "get": {
                "description": "获取指定用户信息",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UmsAdminController"
                ],
                "summary": "获取指定用户信息",
                "parameters": [
                    {
                        "description": "请求信息",
                        "name": "Request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ums_admin.getRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/code.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/ums_admin.getResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
        "/api/admin/{id}": {
            "delete": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "删除管理员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.admin"
                ],
                "summary": "删除管理员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hashId",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.deleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
        "/api/tool/data/mysql": {
            "post": {
                "security": [
                    {
                        "LoginToken": []
                    }
                ],
                "description": "执行 SQL 语句",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API.tool"
                ],
                "summary": "执行 SQL 语句",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据库名称",
                        "name": "db_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "数据表名称",
                        "name": "table_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "SQL 语句",
                        "name": "sql",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tool.searchMySQLResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/code.Failure"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.deleteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "主键ID",
                    "type": "integer"
                }
            }
        },
...
        "code.Failure": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "业务返回数据"
                },
                "message": {
                    "description": "描述信息",
                    "type": "string"
                }
            }
        },
        "code.Success": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务码",
                    "type": "integer"
                },
                "data": {
                    "description": "业务返回数据"
                },
                "message": {
                    "description": "描述信息",
                    "type": "string"
                }
            }
        },
...
        "tool.searchMySQLResponse": {
            "type": "object",
            "properties": {
                "cols": {
                    "description": "查询后的行",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cols_info": {
                    "description": "行信息",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/tool.tableColumn"
                    }
                },
                "list": {
                    "description": "查询后的数据",
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                }
            }
        },
...
        "tool.tableColumn": {
            "type": "object",
            "properties": {
                "column_comment": {
                    "description": "字段注释",
                    "type": "string"
                },
                "column_name": {
                    "description": "字段名称",
                    "type": "string"
                }
            }
        },
...
        "ums_admin.getRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "用户ID",
                    "type": "integer"
                }
            }
        },
        "ums_admin.getResponse": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "loginTime": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "ums_admin.loginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "ums_admin.loginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "tokenHead": {
                    "type": "string"
                }
            }
        },
...
        "ums_admin.registerRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "ums_admin.registerResponse": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "icon": {
                    "description": "头像",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "loginTime": {
                    "description": "最后登录时间",
                    "type": "string"
                },
                "nickName": {
                    "description": "昵称",
                    "type": "string"
                },
                "note": {
                    "description": "备注信息",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "status": {
                    "description": "帐号启用状态：0-\u003e禁用；1-\u003e启用",
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "LoginToken": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}